Use local tor client to connect to destination
Integrate lib stem to start a local installed tor client. Configuration
is within plugin.
Small tor client package have to be installed and path have to be set to
tor executable.
Also it should be possible to connect to a tor client on different
machine.
Currently installed tor as service is not supported (writing this text
this is commin to my mind).
Hope you like it. But this is nothing for one click users!