Added stream type setting
Getting to be difficult to choose which stream to use.  I find HLS
(m3u8) pauses and seeks better especially on Android, but some users
have problems with HLS. So providing preffence. Currently used in
pbskids, but want to look at HLS in the cw since I need to pause
episodes there,